+++ src/components/ids-dropdown/ids-dropdown.ts
@@ -25,7 +25,6 @@ import './ids-dropdown-list';
 import '../ids-list-box/ids-list-box';
 import '../ids-text/ids-text';
 import '../ids-icon/ids-icon';
-import { IdsPopupElementRef } from '../ids-popup/ids-popup-attributes';
 
 import type IdsDropdownList from './ids-dropdown-list';
 import type IdsTriggerButton from '../ids-trigger-field/ids-trigger-button';
@@ -583,14 +582,9 @@ export default class IdsDropdown extends Base {
 
     this.dropdownList.removeTriggerEvents();
     this.dropdownList.appendToTargetParent();
-    this.dropdownList.popupOpenEventsTarget = (this.list ? this : this.container as IdsPopupElementRef);
-    this.dropdownList.onOutsideClick = (e: Event) => {
-      if (this.dropdownList) {
-        if (!e.composedPath()?.includes(this.dropdownList)) {
-          this.close(true);
-        }
-      }
-    };
+    if (this.dropdownList?.popup) {
+      this.dropdownList.popup.onOutsideClick = this.onOutsideClick.bind(this);
+    }
     this.dropdownList.onTriggerClick = () => {
       if (this.labelClicked) {
         this.labelClicked = false;
@@ -607,19 +601,26 @@ export default class IdsDropdown extends Base {
 
     this.dropdownList.setAttribute(attributes.TARGET, `#${targetElemId}`);
     this.dropdownList.setAttribute(attributes.TRIGGER_ELEM, `#${triggerElemId}`);
-    this.dropdownList.popup.alignTarget = this.input?.fieldContainer || this.dropdownList || this;
-
-    this.dropdownList.popupOpenEventsTarget = document.body;
+    if (this.dropdownList?.popup) {
+      this.dropdownList.popup.alignTarget = this.input?.fieldContainer || this.dropdownList || this;
+    }
 
     // Configure inner IdsPopup
     const isRTL = this.localeAPI.isRTL();
-    if (this.locale && isRTL) {
-      this.dropdownList.popup?.setAttribute(attributes.ALIGN, `bottom, ${isRTL || ['lg', 'full'].includes(this.size) ? 'right' : 'left'}`);
-    }
+    this.dropdownList.popup.align = `bottom, ${isRTL || ['lg', 'full'].includes(this.size) ? 'right' : 'left'}`;
+    this.dropdownList.popup.alignEdge = 'bottom';
 
     if (this.input) this.dropdownList.value = this.input.value;
 
     if (this.#isMultiSelect) this.dropdownList.isMultiSelect = true;
+  }
+
+  onOutsideClick(e: Event) {
+    if (this.dropdownList) {
+      if (!e.composedPath()?.includes(this.dropdownList)) {
+        this.close(true);
+      }
+    }
   }
 
   /**
@@ -645,6 +646,7 @@ export default class IdsDropdown extends Base {
     }
 
     // Open the Dropdown List
+    this.configurePopup();
     this.dropdownList.setAttribute(htmlAttributes.ARIA_EXPANDED, 'true');
     this.dropdownList.removeAttribute(attributes.TABINDEX);
     this.dropdownList.show();