+++ src/components/ids-modal/ids-modal.ts
@@ -12,7 +12,7 @@ import IdsElement from '../../core/ids-element';
 
 import { setBooleanAttr } from '../../utils/ids-attribute-utils/ids-attribute-utils';
 import { stringToBool } from '../../utils/ids-string-utils/ids-string-utils';
-import { toggleScrollbar, waitForTransitionEnd } from '../../utils/ids-dom-utils/ids-dom-utils';
+import { toggleScrollbar, waitForTransitionEnd, getClosest } from '../../utils/ids-dom-utils/ids-dom-utils';
 import { cssTransitionTimeout } from '../../utils/ids-timer-utils/ids-timer-utils';
 
 import '../ids-popup/ids-popup';
@@ -110,6 +110,9 @@ export default class IdsModal extends Base {
       this.popup.setAttribute(attributes.TYPE, 'modal');
       this.popup.setAttribute(attributes.ANIMATED, 'true');
       this.popup.setAttribute(attributes.ANIMATION_STYLE, 'scale-in');
+      this.popup.onOutsideClick = this.onOutsideClick.bind(this);
+      this.popup.addOpenEvents = this.addOpenEvents.bind(this);
+      this.popup.removeOpenEvents = this.removeOpenEvents.bind(this);
     }
 
     // Update ARIA / Sets up the label
@@ -509,7 +512,9 @@ export default class IdsModal extends Base {
 
     // Focus the correct element
     this.capturesFocus = true;
-    this.setFocus('last');
+    if (this.autoFocus) {
+      this.setFocus(this.#getFocusableElementIndex());
+    }
 
     this.addOpenEvents();
     this.triggerEvent('show', this, {
@@ -660,8 +665,8 @@ export default class IdsModal extends Base {
    */
   #setFocusIfVisible = async () => {
     this.visible = this.getAttribute('visible');
-    if (this.visible) {
-      this.setFocus('last');
+    if (this.visible && this.autoFocus) {
+      this.setFocus(this.#getFocusableElementIndex());
     }
   };
 
@@ -761,5 +766,12 @@ export default class IdsModal extends Base {
     const closeButton = this.closeButton;
     this.offEvent('click.modal-close', closeButton);
     closeButton?.remove();
+  }
+
+  /**
+   * @returns {number} the index of the first focusable element in the modal content skipping the toolbar
+   */
+  #getFocusableElementIndex(): number {
+    return this.focusableElements?.findIndex((item: HTMLElement) => !getClosest(item, 'ids-toolbar')) || 0;
   }
 }