+++ src/core/ids-attributes.ts
@@ -34,6 +34,7 @@ export const attributes = {
   AUTO_FILL: 'auto-fill',
   AUTO_FIT: 'auto-fit',
   AUTO_FLOW: 'auto-flow',
+  AUTO_FOCUS: 'auto-focus',
   AUTO_HEIGHT: 'auto-height',
   AUTO_ROWS: 'auto-rows',
   AUTO_START: 'auto-start',
@@ -109,6 +110,7 @@ export const attributes = {
   COMPLETED_LABEL: 'completed-label',
   COMPONENT_NAME: 'component-name',
   CONDITION: 'condition',
+  CONFIRM: 'confirm',
   CONTENT_ALIGN: 'content-align',
   COPYRIGHT_YEAR: 'copyright-year',
   COUNT: 'count',
@@ -571,7 +573,7 @@ export const prefix = {
   PREFIX: 'ids'
 };
 
-export const version = '1.0.0-beta.19';
+export const version = '1.0.0-beta.20';
 
 export enum IdsDirection {
   Up = 'up',